+++ ui/pages/Project/selectors.js
@@ -22,21 +22,20 @@ import {
   getMmeResultsByGuid, getMmeSubmissionsByGuid, getHasActiveSearchableSampleByFamily, getSelectableTagTypesByProject,
   getVariantTagsByGuid, getUserOptionsByUsername, getSamplesByFamily, getNotesByFamilyType,
   getVariantTagNotesByFamilyVariants, getPhenotypeGeneScoresByIndividual,
-  getRnaSeqDataByIndividual,
+  getRnaSeqDataByIndividual, familyPassesFilters,
 } from 'redux/selectors'
 
 import {
   SORT_BY_FAMILY_NAME,
   CASE_REVIEW_STATUS_OPTIONS,
   CASE_REVIEW_FILTER_LOOKUP,
-  FAMILY_FILTER_LOOKUP,
   FAMILY_SORT_OPTIONS,
   FAMILY_EXPORT_DATA,
   CASE_REVIEW_FAMILY_EXPORT_DATA,
   CASE_REVIEW_TABLE_NAME,
   CASE_REVIEW_INDIVIDUAL_EXPORT_DATA,
   SAMPLE_EXPORT_DATA,
-  CATEGORY_FAMILY_FILTERS,
+  PROJECT_CATEGORY_FAMILY_FILTERS,
 } from './constants'
 
 const FAMILY_SORT_LOOKUP = FAMILY_SORT_OPTIONS.reduce(
@@ -361,84 +360,18 @@ const getFamilyAnalysers = createSelector(
 export const getFamiliesFilterOptionsByCategory = createSelector(
   getFamilyAnalysers,
   analysedByOptions => ({
-    ...CATEGORY_FAMILY_FILTERS,
+    ...PROJECT_CATEGORY_FAMILY_FILTERS,
     [FAMILY_FIELD_ANALYSED_BY]: [
-      ...CATEGORY_FAMILY_FILTERS[FAMILY_FIELD_ANALYSED_BY],
+      ...PROJECT_CATEGORY_FAMILY_FILTERS[FAMILY_FIELD_ANALYSED_BY],
       ...[...analysedByOptions].map(analysedBy => ({ value: analysedBy, category: 'Analysed By' })),
     ],
   }),
 )
 
-const ANALYSED_BY_FILTER_LOOKUP = Object.values(CATEGORY_FAMILY_FILTERS).reduce(
-  (acc, options) => {
-    options.forEach((opt) => {
-      acc[opt.value] = opt.analysedByFilter
-    })
-    return acc
-  }, {},
-)
-
-const NO_ANALYSED_BY_FIELDS = Object.values(CATEGORY_FAMILY_FILTERS).reduce(
-  (acc, options) => {
-    options.filter(opt => opt.requireNoAnalysedBy).forEach((opt) => {
-      acc.add(opt.value)
-    })
-    return acc
-  }, new Set(),
-)
-
-const ANALYSED_BY_CATEGORY_OPTION_LOOKUP = CATEGORY_FAMILY_FILTERS[FAMILY_FIELD_ANALYSED_BY].reduce(
-  (acc, { value, category }) => ({ ...acc, [value]: category || 'Analysed By' }), {},
-)
-
-const isAnalysedBy = (family, analysedByFilter, user, analysedByOptions) => {
-  let requireNoAnalysedBy = false
-  const analsedByGroups = Object.values(analysedByFilter.reduce(
-    (acc, val) => {
-      const optFilter = analysedByOptions?.has(val) ? ({ createdBy }) => createdBy === val :
-        ANALYSED_BY_FILTER_LOOKUP[val]
-      if (optFilter) {
-        const category = ANALYSED_BY_CATEGORY_OPTION_LOOKUP[val]
-        if (!acc[category]) {
-          acc[category] = []
-        }
-        acc[category].push(optFilter)
-      }
-      if (NO_ANALYSED_BY_FIELDS.has(val)) {
-        requireNoAnalysedBy = true
-      }
-      return acc
-    }, {},
-  ))
-  if (!analsedByGroups.length) {
-    return true
-  }
-  const filteredAnalysedBy = analsedByGroups.reduce(
-    (acc, filterGroup) => acc.filter(analysedBy => filterGroup.some(f => f(analysedBy, user))),
-    family.analysedBy,
-  )
-  return requireNoAnalysedBy ? filteredAnalysedBy.length === 0 : filteredAnalysedBy.length > 0
-}
-
 export const getFamiliesTableFilters = createSelector(
   getFamiliesTableFiltersByProject,
   getProjectGuid,
   (familyTableFiltersByProject, projectGuid) => (familyTableFiltersByProject || {})[projectGuid],
-)
-
-const familyPassesFilters = createSelector(
-  getIndividualsByGuid,
-  getUser,
-  getSamplesByFamily,
-  (individualsByGuid, user, samplesByFamily) => (family, groupedFilters, analysedByOptions) => {
-    if (groupedFilters.analysedBy && !isAnalysedBy(family, groupedFilters.analysedBy, user, analysedByOptions)) {
-      return false
-    }
-    return Object.values(groupedFilters).every((groupVals) => {
-      const filters = (groupVals || []).map(val => FAMILY_FILTER_LOOKUP[val]).filter(val => val)
-      return !filters.length || filters.some(filter => filter(family, individualsByGuid, user, samplesByFamily))
-    })
-  },
 )
 
 const familyPassesTableFilters = createSelector(
